Update Candida-auris-CladeI-1.tsv
Email from Rory Welsh 2021-05-05

Hi Lee, 

Could we update the C. auris benchmark dataset in the GMI github dataset folder with the attached file? The publication announcing this benchmark dataset is live and AMD requested content to promote the recent publication (see below), which reminded me about the updated file. During the publication process we decided to drop one of the isolates and produced a new tree with the final attached dataset. The attached txt file has that isolated and row deleted, an updated link to the final tree, and includes the publication’s PubMed ID. 

Regards,
Rory